Combined the Altera targets into one project
The purpose of this was to eliminate the redundancy of the AHDL and
Verilog files between the three Altera targets (DE2-115, DE0-Nano and
BeMicroCV). All three targets and their necessary source files are now
in the Altera directory, and the output_files subdirectory is used for
all output. To compile one of the targets, load the Altera/P8X32A.qpf
project file and select the Revision for your target. Conversion of the
programming files is the same except the file names are different.
I accomplished this by copying all the files from the original DE2
directory, and copying and renaming the original .qsf files. That means
the configuration for each target is the same as it used to be when each
target had its own project file. The .cof files were also changed to
convert the appropriate output files (instead of "top.sof). I modified
the hub_mem.v module so that the inclusion of the character ROM depends
on a new setting in the DE0-Nano revision but I'm planning on using a
better system in the future. The inversion of the BeMicro LEDs is back
on when the Altera tree is used (i.e. LED on=cog not in use, LED off=cog
in use). This is a temporary measure until I figure out how to
parameterize things in a good way.
I didn't delete the old directories yet, in order to make it possible to
pull this change and compare files between directories. Unfortunately
Git doesn't know how to branch files between directories (only between
branches) so there's no way for me to force the new files to have all
the old history, and I'm not sure how good Git is at detecting the
copied files, especially because there were already duplicates in
existence.